The size ratchet refused the new tool, and it was right to
qaSection.ts sat at EXACTLY its test_file_sizes.py pin — 1,334 lines, zero
headroom — and the wall-joins tool is 80 lines. The full backend suite came back
667/668 with test_file_sizes the only failure: "no file under an extraction
ratchet has grown".

I checked this constraint for app.ts and not for the file I was actually editing.
Last sprint the skylight went into envelopeSection.ts specifically to avoid
paying for a tool by unwinding an extraction — the same reasoning, applied to the
wrong file, because I looked up the rule instead of looking up the pin.

The remedy is the one the pin's own note records from the R41 yaw fit: "the panel
was extracted instead". Both REPAIR controls move to tools/repairPanel.ts:

  - extracting only the new tool costs 2 net lines here (an import and a call)
    and would still have been over the pin. Taking the cleanup tool along is what
    PAYS for the addition rather than merely relocating it.
  - and the two belong together on the merits. They are the only controls in the
    app that repair the model rather than report on it, and they are the two the
    roadmap's "diagnosis ships, the repair does not" list got wrong in OPPOSITE
    directions — cleanup's repair had shipped all along; wall joins had neither
    half on screen.

qaSection.ts 1,334 -> 1,305; the pin is re-tightened to the exact new count, not
left with slack, per the note above it: "13 lines of slack is 13 lines the next
addition spends without anyone deciding to."

AND THE MOVE BROKE A SECOND GATE, WHICH IS THE INTERESTING PART. toolsSplit.test.ts
reads SECTION_SOURCES as text and splits the CONCATENATION on `const b = section("key"`,
running each body to the next marker. Every other file in that list opens its own
section() and so labels itself; repairPanel.ts does not — it exports two buttons
qaSection appends. Appended at the END of the list its text was attributed to
`authoring`, and the cleanup control read as having CHANGED SIDES. Sitting directly
after qaSection.ts it reads as `qa`, which is where the buttons mount.

*A list that looks like a set can still be an order.* That is now stated at the
entry, because nothing else in the file would tell the next person.

web 208 files / 2108 tests; typecheck clean; eslint clean on all three files;
test_file_sizes passes including its narrative-chain check (1334 -> 1305 recorded,
no discontinuity).
