+++ services/api/test_file_sizes.py
@@ -134,7 +134,7 @@ PER_FILE = {
     # not to raise - the same call the /finance extraction made earlier the same day. Re-tightened
     # to the new exact count rather than left at 1_373, because 13 lines of slack is 13 lines the
     # next addition spends without anyone deciding to.
-    "apps/web/src/viewer/tools/qaSection.ts": 1_334,   # v0.3.1134: blank import gap closed (1_335 -> 1_334) while geometry results now show sourced basis.   # alignment check -> tools/alignmentPanel.ts (1_348 -> 1_335). The ratchet is why: adding the R41 yaw fit inline pushed this file 17 lines over and the gate refused, so the panel was extracted instead and came back three lines.
+    "apps/web/src/viewer/tools/qaSection.ts": 1_305,   # WALL-JOINS: the two REPAIR controls -> tools/repairPanel.ts (1_334 -> 1_305). The ratchet is the reason there is a module at all: this file sat at its pin EXACTLY, with zero headroom, and the wall-joins tool is 80 lines, so the gate refused — the same refusal the alignment note below records. Extracting only the new tool would have cost 2 net lines (an import and a call) and still been over, so the cleanup tool came along, which is what PAYS for the addition rather than merely relocating it. The two belong together on the merits: they are the only controls in the app that REPAIR the model rather than report on it, and they are the two the roadmap's "diagnosis ships, the repair does not" list got wrong in opposite directions.   # v0.3.1134: blank import gap closed (1_335 -> 1_334) while geometry results now show sourced basis.   # alignment check -> tools/alignmentPanel.ts (1_348 -> 1_335). The ratchet is why: adding the R41 yaw fit inline pushed this file 17 lines over and the gate refused, so the panel was extracted instead and came back three lines.
     # Pinned at its EXACT measured size on 2026-08-09, BEFORE it needs splitting — the third-largest
     # hand-written file in the tree and the only one of that size with no ratchet. The qaSection entry
     # above records why the timing matters: "a ratchet added at the point of pain only ratifies the
